Revert "refactor: /tests/integration/650.command.envelope.test.cjs (#5919)"
This reverts commit e3d2348e391b0420c8cff853dff062ef2d954a2e.
