+++ tests/integration/snapshots/650.command.envelope.test.cjs.md
@@ -1,0 +1,23 @@
+# Snapshot report for `tests/integration/650.command.envelope.test.cjs`
+
+The actual snapshot is saved in `650.command.envelope.test.cjs.snap`.
+
+Generated by [AVA](https://avajs.dev).
+
+## env:clone should return success message (mongo to envelope)
+
+> Snapshot 1
+
+    'Successfully cloned environment variables from site-name-a to site-name-b'
+
+## env:clone should return success message (envelope to mongo)
+
+> Snapshot 1
+
+    'Successfully cloned environment variables from site-name-a to site-name-b'
+
+## env:clone should return success message (envelope to envelope)
+
+> Snapshot 1
+
+    'Successfully cloned environment variables from site-name-a to site-name-b'
